+++ hastexo/tasks.py
@@ -19,6 +19,7 @@ from tenacity import (
     before_sleep_log,
 )
 
+from .celery import app
 from .models import Stack
 from .provider import Provider, ProviderException
 from .common import (
@@ -609,6 +610,9 @@ class LaunchStackTask(HastexoTask):
         return check_data
 
 
+LaunchStackTask = app.register_task(LaunchStackTask())
+
+
 class SuspendStackTask(HastexoTask):
     """
     Suspends a stack.
@@ -686,6 +690,9 @@ class SuspendStackTask(HastexoTask):
                          provider_stack["status"])
 
         return provider_stack["status"]
+
+
+SuspendStackTask = app.register_task(SuspendStackTask())
 
 
 class DeleteStackTask(HastexoTask):
@@ -805,6 +812,9 @@ class DeleteStackTask(HastexoTask):
         return provider_stack["status"]
 
 
+DeleteStackTask = app.register_task(DeleteStackTask())
+
+
 class CheckStudentProgressTask(HastexoTask):
     """
     Check student progress by running a set of scripts via SSH.
@@ -864,3 +874,6 @@ class CheckStudentProgressTask(HastexoTask):
             'total': len(self.tests),
             'errors': errors
         }
+
+
+CheckStudentProgressTask = app.register_task(CheckStudentProgressTask())